Implement auto-check of ray/pyopencl for NLPAR.
Signed-off by: David Rowenhorst <david.rowenhorst@nrl.navy.mil>
