+++ pyebsdindex/opencl/nlpar_cl.py
@@ -1,18 +1,45 @@
+# This software was developed by employees of the US Naval Research Laboratory (NRL), an
+# agency of the Federal Government. Pursuant to title 17 section 105 of the United States
+# Code, works of NRL employees are not subject to copyright protection, and this software
+# is in the public domain. PyEBSDIndex is an experimental system. NRL assumes no
+# responsibility whatsoever for its use by other parties, and makes no guarantees,
+# expressed or implied, about its quality, reliability, or any other characteristic. We
+# would appreciate acknowledgment if the software is used. To the extent that NRL may hold
+# copyright in countries other than the United States, you are hereby granted the
+# non-exclusive irrevocable and unconditional right to print, publish, prepare derivative
+# works and distribute this software, in any medium, or authorize others to do so on your
+# behalf, on a royalty-free basis throughout the world. You may improve, modify, and
+# create derivative works of the software or any portion of the software, and you may copy
+# and distribute such modifications or works. Modified works should carry a notice stating
+# that you changed the software and should note the date and nature of any such change.
+# Please explicitly acknowledge the US Naval Research Laboratory as the original source.
+# This software can be redistributed and/or modified freely provided that any derivative
+# works bear some notice that they are derived from it, and any modified versions bear
+# some notice that they have been modified.
+#
+# Author: David Rowenhorst;
+# The US Naval Research Laboratory Date: 22 May 2024
+
+# For more info see
+# Patrick T. Brewick, Stuart I. Wright, David J. Rowenhorst. Ultramicroscopy, 200:50–61, May 2019.
+
+
+
 import os
 from timeit import default_timer as timer
 import numpy as np
 import pyopencl as cl
-import pyopencl.array as cl_array
 
-import numba
+
+
 import scipy.optimize as sp_opt
-from pyebsdindex import nlpar
+from pyebsdindex import nlpar_cpu
 from pyebsdindex.opencl import openclparam
 from time import time as timer
-import scipy
-class NLPAR(nlpar.NLPAR):
+
+class NLPAR(nlpar_cpu.NLPAR):
   def __init__( self, filename=None, **kwargs):
-    nlpar.NLPAR.__init__(self, filename=filename, **kwargs)
+    nlpar_cpu.NLPAR.__init__(self, filename=filename, **kwargs)
     self.useCPU = False
 
 
@@ -30,6 +57,12 @@ class NLPAR(nlpar.NLPAR):
 
   def calcsigma(self,nn=1, saturation_protect=True,automask=True, return_nndist=False, **kwargs):
     self.sigmann = nn
+    if self.sigmann > 7:
+      print("Sigma optimization search limited to a search radius <= 7")
+      print("The search radius has been clipped to 7")
+      nn = 7
+      self.sigmann = nn
+
     sig =  self.calcsigma_cl(nn=nn,
                             saturation_protect=saturation_protect,
                             automask=automask, **kwargs)
@@ -38,14 +71,14 @@ class NLPAR(nlpar.NLPAR):
     else:
       return sig[0]
   def opt_lambda_cpu(self, **kwargs):
-    return nlpar.NLPAR.opt_lambda(self, **kwargs)
+    return nlpar_cpu.NLPAR.opt_lambda(self, **kwargs)
 
   def calcnlpar_cpu(self, **kwargs):
-    return nlpar.NLPAR.calcnlpar(self, **kwargs)
+    return nlpar_cpu.NLPAR.calcnlpar(self, **kwargs)
 
   def calcsigma_cpu(self,nn=1, saturation_protect=True,automask=True, **kwargs):
-    return nlpar.NLPAR.calcsigma(self, nn=nn,
-                            saturation_protect=saturation_protect,automask=automask, **kwargs)
+    return nlpar_cpu.NLPAR.calcsigma(self, nn=nn,
+                                     saturation_protect=saturation_protect, automask=automask, **kwargs)
 
   def opt_lambda_cl(self, saturation_protect=True, automask=True, backsub=False,
                  target_weights=[0.5, 0.34, 0.25], dthresh=0.0, autoupdate=True, **kwargs):
@@ -101,6 +134,12 @@ class NLPAR(nlpar.NLPAR):
 
 
   def calcsigma_cl(self,nn=1,saturation_protect=True,automask=True, normalize_d=False, gpu_id = None, **kwargs):
+    self.sigmann = nn
+    if self.sigmann > 7:
+      print("Sigma optimization search limited to a search radius <= 7")
+      print("The search radius has been clipped to 7")
+      nn = 7
+      self.sigmann = nn
 
     if gpu_id is None:
       clparams = openclparam.OpenClParam()
@@ -258,7 +297,7 @@ class NLPAR(nlpar.NLPAR):
     self.sigma = sigma
     return sigma, dist, countnn
 
-  def calcnlpar_cl(self, chunksize=0, searchradius=None, lam = None, dthresh = None, saturation_protect=True, automask=True,
+  def calcnlpar_cl(self, searchradius=None, lam = None, dthresh = None, saturation_protect=True, automask=True,
                    filename=None, fileout=None, reset_sigma=False, backsub = False, rescale = False,
                    gpu_id = None, verbose=2, **kwargs):
 
@@ -273,7 +312,11 @@ class NLPAR(nlpar.NLPAR):
     if searchradius is not None:
       self.searchradius = searchradius
 
-
+    if self.searchradius > 10:
+      print("NLPAR on GPU is limited to a search radius <= 10")
+      print("The search radius has been clipped to 10")
+      searchradius = 10
+      self.searchradius = searchradius
 
     lam = np.float32(self.lam)
     dthresh = np.float32(self.dthresh)
