+++ pyebsdindex/opencl/nlpar_clray.py
@@ -1,3 +1,32 @@
+# This software was developed by employees of the US Naval Research Laboratory (NRL), an
+# agency of the Federal Government. Pursuant to title 17 section 105 of the United States
+# Code, works of NRL employees are not subject to copyright protection, and this software
+# is in the public domain. PyEBSDIndex is an experimental system. NRL assumes no
+# responsibility whatsoever for its use by other parties, and makes no guarantees,
+# expressed or implied, about its quality, reliability, or any other characteristic. We
+# would appreciate acknowledgment if the software is used. To the extent that NRL may hold
+# copyright in countries other than the United States, you are hereby granted the
+# non-exclusive irrevocable and unconditional right to print, publish, prepare derivative
+# works and distribute this software, in any medium, or authorize others to do so on your
+# behalf, on a royalty-free basis throughout the world. You may improve, modify, and
+# create derivative works of the software or any portion of the software, and you may copy
+# and distribute such modifications or works. Modified works should carry a notice stating
+# that you changed the software and should note the date and nature of any such change.
+# Please explicitly acknowledge the US Naval Research Laboratory as the original source.
+# This software can be redistributed and/or modified freely provided that any derivative
+# works bear some notice that they are derived from it, and any modified versions bear
+# some notice that they have been modified.
+#
+# Author: David Rowenhorst;
+# The US Naval Research Laboratory Date: 22 May 2024
+
+# For more info see
+# Patrick T. Brewick, Stuart I. Wright, David J. Rowenhorst. Ultramicroscopy, 200:50–61, May 2019.
+
+"""Non-local pattern averaging and re-indexing (NLPAR)."""
+
+
+
 import os, sys, platform
 import logging
 from timeit import default_timer as timer
@@ -6,7 +35,7 @@ import pyopencl as cl
 
 import ray
 
-#from pyebsdindex import nlpar
+
 from pyebsdindex.opencl import openclparam, nlpar_cl
 from time import time as timer
 
@@ -25,7 +54,12 @@ class NLPAR(nlpar_cl.NLPAR):
     return self.calcnlpar_clray(**kwargs)
 
   def calcsigma(self, nn=1, saturation_protect=True, automask=True, return_nndist=False, **kwargs):
-    self.sigmann = nn
+    if self.sigmann > 7:
+      print("Sigma optimization search limited to a search radius <= 7")
+      print("The search radius has been clipped to 7")
+      nn = 7
+      self.sigmann = nn
+
     sig = self.calcsigma_clray(nn=nn,
                             saturation_protect=saturation_protect,
                             automask=automask, **kwargs)
@@ -45,6 +79,12 @@ class NLPAR(nlpar_cl.NLPAR):
                       gpu_id = None, verbose=2, **kwargs):
     self.patternfile = self.getinfileobj()
     self.sigmann = nn
+
+    if self.sigmann > 7:
+      print("Sigma optimization search limited to a search radius <= 7")
+      print("The search radius has been clipped to 7")
+      nn = 7
+      self.sigmann = nn
 
     if gpu_id is None:
       clparams = openclparam.OpenClParam()
@@ -218,6 +258,7 @@ class NLPAR(nlpar_cl.NLPAR):
 
   def _sigmachunkcalc_cl(self, data, calclim, clparams=None, saturation_protect=True):
     nn = self.sigmann
+
     data = np.ascontiguousarray(data)
     ctx = clparams.ctx
     prg = clparams.prg
@@ -343,6 +384,12 @@ class NLPAR(nlpar_cl.NLPAR):
       self.dthresh = 0.0
 
     if searchradius is not None:
+      self.searchradius = searchradius
+
+    if self.searchradius > 10:
+      print("NLPAR on GPU is limited to a search radius <= 10")
+      print("The search radius has been clipped to 10")
+      searchradius = 10
       self.searchradius = searchradius
 
     lam = np.float32(self.lam)
